Renamed TocDocFxCreation to DocFxTocGenerator. Aligned namespaces
