+++ THIRD-PARTY-NOTICES.TXT
@@ -1,6 +1,8 @@
 Most of those companion tools has been written jointly with ZF under a proper MIT license. 
+
 The MIT License (MIT)
 Copyright © 2020, ZF and contributors
+
 Permission is hereby granted, free of charge, to any person obtaining a copy
 of this software and associated documentation files (the “Software”), to deal
 in the Software without restriction, including without limitation the rights
