+++ src/DocFxTocGenerator/Domain/CommandlineOptions.cs
@@ -1,7 +1,7 @@
 ﻿// Licensed to DocFX Companion Tools and contributors under one or more agreements.
 // DocFX Companion Tools and contributors licenses this file to you under the MIT license.
 
-namespace TocDocFxCreation.Domain
+namespace DocFxTocGenerator.Domain
 {
     using CommandLine;
