+++ src/DocFxTocGenerator/Helpers/MessageHelper.cs
@@ -1,10 +1,10 @@
 ﻿// Licensed to DocFX Companion Tools and contributors under one or more agreements.
 // DocFX Companion Tools and contributors licenses this file to you under the MIT license.
 
-namespace TocDocFxCreation.Helpers
+namespace DocFxTocGenerator.Helpers
 {
     using System;
-    using TocDocFxCreation.Domain;
+    using DocFxTocGenerator.Domain;
 
     /// <summary>
     /// Helper methods to write messages to the console.
