+++ src/DocFxTocGenerator/Properties/launchSettings.json
@@ -1,6 +1,6 @@
 {
   "profiles": {
-    "TocDocFxCreation": {
+    "DocFxTocGenerator": {
       "commandName": "Project",
       "commandLineArgs": "-d .\\docs -o .\\docs -v -s -i"
     }
